+++ Preventanyl/Preventanyl/VCMapView.swift
@@ -107,4 +107,17 @@ extension FirstViewController: MKMapViewDelegate {
         mapItem.openInMaps(launchOptions: options)
     }
     
+    func addMapTrackingButton(){
+        let button   = UIButton(type: UIButtonType.custom) as UIButton
+        button.frame = CGRect(origin: CGPoint(x:5, y: 25), size: CGSize(width: 35, height: 35))
+        button.setImage(#imageLiteral(resourceName: "first"), for: .normal)
+        button.backgroundColor = .clear
+        button.addTarget(self, action: #selector(FirstViewController.centerMapOnUserButtonClicked), for:.touchUpInside)
+        MapView.addSubview(button)
+    }
+    
+    @objc func centerMapOnUserButtonClicked() {
+        MapView.setUserTrackingMode(MKUserTrackingMode.follow, animated: true)
+    }
+    
 }